# internal/docs/api/swagger.yaml
definitions:
...
  CreateStationObservationReq:
    properties:
      hi:
        type: number
      mslp:
        type: number
      pres:
        type: number
      qc_level:
        type: integer
      rh:
        type: number
      rr:
        type: number
      srad:
        type: number
      station_id:
        type: integer
      td:
        type: number
      temp:
        type: number
      timestamp:
        type: string
      wchill:
        type: number
      wdir:
        type: number
      wspd:
        type: number
      wspdx:
        type: number
    type: object
...
  LatestObservation:
    properties:
      address:
        type: string
      elevation:
        type: number
      id:
        type: integer
      lat:
        type: number
      lon:
        type: number
      name:
        type: string
      obs:
        $ref: '#/definitions/handlers.latestObsRes'
    type: object
...
  PaginatedStationObservations:
    type: object
...
  StationObservation:
    properties:
      hi:
        type: number
      id:
        type: integer
      mslp:
        type: number
      pres:
        type: number
      qc_level:
        type: integer
      rh:
        type: number
      rr:
        type: number
      srad:
        type: number
      station_id:
        type: integer
      td:
        type: number
      temp:
        type: number
      timestamp:
        type: string
      wchill:
        type: number
      wdir:
        type: number
      wspd:
        type: number
      wspdx:
        type: number
    type: object
...
  UpdateStationObservationParams:
    properties:
      hi:
        type: number
      id:
        type: integer
      mslp:
        type: number
      pres:
        type: number
      qc_level:
        type: integer
      rh:
        type: number
      rr:
        type: number
      srad:
        type: number
      station_id:
        type: integer
      td:
        type: number
      temp:
        type: number
      timestamp:
        type: string
      wchill:
        type: number
      wdir:
        type: number
      wspd:
        type: number
      wspdx:
        type: number
    type: object
...
  handlers.latestObsRes:
    properties:
      gust:
        type: number
      gust_timestamp:
        type: string
      mslp:
        type: number
      rain:
        type: number
      rain_accum:
        type: number
      rh:
        type: number
      srad:
        type: number
      temp:
        type: number
      timestamp:
        type: string
      tn:
        type: number
      tn_timestamp:
        type: string
      tx:
        type: number
      tx_timestamp:
        type: string
      wdir:
        type: number
      wspd:
        type: number
    type: object
...
info:
  contact:
    email: emiliogozo@proton.me
    name: Emilio Gozo
  description: Panahon API.
  title: Panahon API
  version: "1.0"
paths:
...
  /observations:
    get:
      parameters:
      - in: query
        name: end_date
        type: string
      - description: page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: limit
        in: query
        maximum: 50000
        minimum: 50
        name: per_page
        type: integer
      - in: query
        name: start_date
        type: string
      - in: query
        name: station_ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PaginatedStationObservations'
      summary: list station observation
      tags:
      - observations
...
  /stations/{station_id}/observations:
    get:
      consumes:
      - application/json
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      - in: query
        name: end_date
        type: string
      - description: page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: limit
        in: query
        minimum: 1
        name: per_page
        type: integer
      - in: query
        name: start_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PaginatedStationObservations'
      summary: List station observations
      tags:
      - observations
    post:
      consumes:
      - application/json
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      - description: Create station observation parameters
        in: body
        name: stnObs
        required: true
        schema:
          $ref: '#/definitions/CreateStationObservationReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/StationObservation'
      security:
      - BearerAuth: []
      summary: Create station observation
      tags:
      - observations
  /stations/{station_id}/observations/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      - description: Station Observation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete station observation
      tags:
      - observations
    get:
      consumes:
      - application/json
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      - description: Station Observation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/StationObservation'
      summary: Get station observation
      tags:
      - observations
    put:
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      - description: Station Observation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update station observation parameters
        in: body
        name: stnObs
        required: true
        schema:
          $ref: '#/definitions/UpdateStationObservationParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/StationObservation'
      security:
      - BearerAuth: []
      summary: Update station observation
      tags:
      - observations
  /stations/{station_id}/observations/latest:
    get:
      consumes:
      - application/json
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LatestObservation'
      summary: Get latest station observation
      tags:
      - observations
  /stations/nearest/observations/latest:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: pt
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LatestObservation'
      summary: Get nearest latest station observation
      tags:
      - observations
...
securityDefinitions:
  APITokenAuth:
    in: query
    name: api_key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"